refactor: reorganize project structure and update configurations
