+++ gitlab_integration/webhook_listener.py
@@ -3,12 +3,9 @@ import threading
 
 from flask import request, jsonify
 
-from app.gitlab_utils import get_merge_request_id, get_commit_change_file, get_merge_request_changes
 from gitlab_integration.gitlab_fetcher import GitlabMergeRequestFetcher
 from reply_module.reply import Reply
 from review_engine.review_engine import ReviewEngine
-from service.chat_review import review_code_for_mr, review_code_for_add_commit
-from utils.dingding import send_dingtalk_message_by_sign
 from utils.logger import log
 
 
@@ -75,20 +72,6 @@ class WebhookListener:
         """
         处理推送事件
         """
-        project_id = gitlab_payload.get('project')['id']
-        merge_request_id = get_merge_request_id(gitlab_payload.get('ref').split("/")[-1], gitlab_payload.get("project_id"))
-
-
-        if not merge_request_id:
-            send_dingtalk_message_by_sign(
-                f"Project_Name:{gitlab_payload['project']['name']}\n备注：分支 {gitlab_payload.get('ref')} 没有处于open状态的 Merge Request 不进行 Code Review。")
-            return jsonify({'status': f'非存在MR分支,{gitlab_payload}'}), 200
-
-        changed_files = get_commit_change_file(gitlab_payload)
-
-        thread = threading.Thread(target=review_code_for_add_commit,
-                                  args=(project_id, merge_request_id, changed_files, gitlab_payload))
-        thread.start()
 
         return jsonify({'status': 'success'}), 200
